scraper/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pyyaml>=6.0
# Optional: for JS-rendered pages (e.g. Google Sites)
# pip install playwright && playwright install chromium
